+++ CHANGELOG.md
@@ -6,6 +6,9 @@
   - Fixes some cases of wrong culling. 
 
 
+- Added Payback Request Canceling recipe to Mail Service. 
+
+
 - When moving mail recipe ingredients into a Package with JEI - address of the recipe will be remembered and applied when items are packed, consuming an Address Tag.    
 - Packages from Create can be delivered now.
   - Address Tag is still required, regardless of the address on the package. Address Tag can be applied automatically with crafters.